requirements.txt
RPi.GPIO
adafruit-circuitpython-servokit
opencv-python-headless
flask
